requirements.txt
groq>=0.9.0
deepeval>=0.21.0
pytest>=7.4.0
pytest-html>=4.0.0
python-dotenv>=1.0.0
